Improve realm status handling
Allows for a realm to come back online before receiving a link down notification from the previous service. The old handling would have caused the link down notification to have set the realm as offline on the list even if it had already restarted. Not likely to have happened but best to handle it.